Wrap the generate page with a suspense boundary to resolve the deployment error
